api+frontend: contributors NUMERIC→BIGINT cast, TheoremBrowser handles imported rows, K2.5→K2.6 doc sweep
- `/api/contributors` was 500ing on decode because Postgres `SUM(bigint)`
  widens to NUMERIC, which the `ContributorStats.theorems_contributed: i64`
  decoder rejects. Adds an explicit `::BIGINT` cast on all three
  aggregates so the result frame matches the Rust types. This is why the
  leaderboard showed `0 contributors` while `/api/workers` correctly
  reported `1 worker live`.

- TheoremBrowser landing teaser now renders imported PhysLean rows the
  same way ResultCard/TheoremDetail already do: Lean qualifier last
  segment as the headline, canonical statement in a monospace pre, no
  KaTeX attempt when `latex` is null.

- Doc-comment sweep: remaining `Kimi K2.5` references in steerer/mod.rs,
  llm/provider.rs, and PipelineDiagram.tsx switched to `K2.6`. The
  Gradient `SUPPORTED` list still keeps `kimi-k2.5` as a migration
  fallback — that's load-bearing, not stale.

+++ engine/crates/pg/src/query/contributors.rs
@@ -34,9 +34,9 @@ pub async fn list_contributors(db: &impl ConnectionTrait) -> Result<Vec<Contribu
             u.id as user_id,
             u.display_name,
             SPLIT_PART(u.email, '@', 1) as handle,
-            COUNT(DISTINCT w.id) as worker_count,
-            COALESCE(SUM(w.theorems_contributed), 0) as theorems_contributed,
-            COUNT(DISTINCT CASE WHEN w.status = 'active' THEN w.id END) as active_worker_count
+            COUNT(DISTINCT w.id)::BIGINT as worker_count,
+            COALESCE(SUM(w.theorems_contributed), 0)::BIGINT as theorems_contributed,
+            COUNT(DISTINCT CASE WHEN w.status = 'active' THEN w.id END)::BIGINT as active_worker_count
         FROM users u
         INNER JOIN api_keys ak ON ak.user_id = u.id AND ak.kind = 'worker' AND ak.revoked_at IS NULL
         INNER JOIN workers w ON w.id = ak.name
